backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
youtube-transcript-api==0.6.1
openai==1.3.0
transformers==4.35.2
torch>=2.0.0
httpx==0.25.1
python-multipart==0.0.6
pytube==15.0.0 